@@ -1,5 +1,14 @@
 <?php
 
+use App\Http\Controllers\Api\V1\Auth\AuthenticateController;
+use App\Http\Controllers\Api\V1\CommentController;
+use App\Http\Controllers\Api\V1\MediaController;
+use App\Http\Controllers\Api\V1\PostCommentController;
+use App\Http\Controllers\Api\V1\PostController;
+use App\Http\Controllers\Api\V1\PostLikeController;
+use App\Http\Controllers\Api\V1\UserCommentController;
+use App\Http\Controllers\Api\V1\UserController;
+use App\Http\Controllers\Api\V1\UserPostController;
 use Illuminate\Support\Facades\Route;
 
 /*
@@ -13,38 +22,38 @@ use Illuminate\Support\Facades\Route;
 |
 */
 
-Route::prefix('v1')->namespace('Api\V1')->group(function () {
+Route::prefix('v1')->group(function () {
     Route::middleware(['auth:api', 'verified'])->group(function () {
         // Comments
-        Route::apiResource('comments', 'CommentController')->only('destroy');
-        Route::apiResource('posts.comments', 'PostCommentController')->only('store');
+        Route::apiResource('comments', CommentController::class)->only('destroy');
+        Route::apiResource('posts.comments', PostCommentController::class)->only('store');
 
         // Posts
-        Route::apiResource('posts', 'PostController')->only(['update', 'store', 'destroy']);
-        Route::post('/posts/{post}/likes', 'PostLikeController@store')->name('posts.likes.store');
-        Route::delete('/posts/{post}/likes', 'PostLikeController@destroy')->name('posts.likes.destroy');
+        Route::apiResource('posts', PostController::class)->only(['update', 'store', 'destroy']);
+        Route::post('/posts/{post}/likes', [PostLikeController::class, 'store'])->name('posts.likes.store');
+        Route::delete('/posts/{post}/likes', [PostLikeController::class, 'destroy'])->name('posts.likes.destroy');
 
         // Users
-        Route::apiResource('users', 'UserController')->only('update');
+        Route::apiResource('users', UserController::class)->only('update');
 
         // Media
-        Route::apiResource('media', 'MediaController')->only(['store', 'destroy']);
+        Route::apiResource('media', MediaController::class)->only(['store', 'destroy']);
     });
 
-    Route::post('/authenticate', 'Auth\AuthenticateController@authenticate')->name('authenticate');
+    Route::post('/authenticate', [AuthenticateController::class, 'authenticate'])->name('authenticate');
 
     // Comments
-    Route::apiResource('posts.comments', 'PostCommentController')->only('index');
-    Route::apiResource('users.comments', 'UserCommentController')->only('index');
-    Route::apiResource('comments', 'CommentController')->only(['index', 'show']);
+    Route::apiResource('posts.comments', PostCommentController::class)->only('index');
+    Route::apiResource('users.comments', UserCommentController::class)->only('index');
+    Route::apiResource('comments', CommentController::class)->only(['index', 'show']);
 
     // Posts
-    Route::apiResource('posts', 'PostController')->only(['index', 'show']);
-    Route::apiResource('users.posts', 'UserPostController')->only('index');
+    Route::apiResource('posts', PostController::class)->only(['index', 'show']);
+    Route::apiResource('users.posts', UserPostController::class)->only('index');
 
     // Users
-    Route::apiResource('users', 'UserController')->only(['index', 'show']);
+    Route::apiResource('users', UserController::class)->only(['index', 'show']);
 
     // Media
-    Route::apiResource('media', 'MediaController')->only('index');
+    Route::apiResource('media', MediaController::class)->only('index');
 });
