@@ -1,5 +1,9 @@
 <?php
 
+use App\Http\Controllers\NewsletterSubscriptionController;
+use App\Http\Controllers\PostController;
+use App\Http\Controllers\PostFeedController;
+use App\Http\Controllers\UserController;
 use Illuminate\Support\Facades\Route;
 
 /*
@@ -13,9 +17,9 @@ use Illuminate\Support\Facades\Route;
 |
 */
 
-Route::get('/', 'PostController@index')->name('home');
-Route::get('/posts/feed', 'PostFeedController@index')->name('posts.feed');
-Route::resource('posts', 'PostController')->only('show');
-Route::resource('users', 'UserController')->only('show');
+Route::get('/', [PostController::class, 'index'])->name('home');
+Route::get('/posts/feed', [PostFeedController::class, 'index'])->name('posts.feed');
+Route::resource('posts', PostController::class)->only('show');
+Route::resource('users', UserController::class)->only('show');
 
-Route::get('newsletter-subscriptions/unsubscribe', 'NewsletterSubscriptionController@unsubscribe')->name('newsletter-subscriptions.unsubscribe');
+Route::get('newsletter-subscriptions/unsubscribe', [NewsletterSubscriptionController::class, 'unsubscribe'])->name('newsletter-subscriptions.unsubscribe');
