@@ -61,7 +61,6 @@ class RouteServiceProvider extends ServiceProvider
     protected function mapWebRoutes(): void
     {
         Route::middleware('web')
-             ->namespace($this->namespace)
              ->group(base_path('routes/web.php'));
     }
 
@@ -74,7 +73,6 @@ class RouteServiceProvider extends ServiceProvider
     {
         Route::prefix('api')
              ->middleware('api')
-             ->namespace($this->namespace)
              ->group(base_path('routes/api.php'));
     }
 
@@ -87,7 +85,6 @@ class RouteServiceProvider extends ServiceProvider
     {
         Route::prefix('admin')
              ->middleware(['web', 'auth', 'role:admin', 'verified'])
-             ->namespace($this->namespace . '\Admin')
              ->as('admin.')
              ->group(base_path('routes/admin.php'));
     }
@@ -100,7 +97,6 @@ class RouteServiceProvider extends ServiceProvider
     protected function mapAuthRoutes(): void
     {
         Route::middleware('web')
-             ->namespace($this->namespace)
              ->group(base_path('routes/auth.php'));
     }
 
@@ -112,7 +108,7 @@ class RouteServiceProvider extends ServiceProvider
     protected function configureRateLimiting()
     {
         RateLimiter::for('api', function (Request $request) {
-            return Limit::perMinute(60);
+            return Limit::perMinute(60)->by($request->user()?->id ?: $request->ip());
         });
     }
 }
