@@ -4,8 +4,7 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-class AddProviderAndProviderIdToUsers extends Migration
-{
+return new class extends Migration {
     /**
      * Run the migrations.
      *
@@ -32,4 +31,4 @@ class AddProviderAndProviderIdToUsers extends Migration
             $table->dropColumn('provider_id');
         });
     }
-}
+};
