@@ -4,8 +4,7 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-class CreateMediaTableWithMediaLibrary extends Migration
-{
+return new class extends Migration {
     /**
      * Run the migrations.
      */
@@ -46,4 +45,4 @@ class CreateMediaTableWithMediaLibrary extends Migration
 
         Schema::dropIfExists('media');
     }
-}
+};
